Django_Social_Network_App/requirements.txt
# Ensure latest pip, setuptools and wheel for Python 3.13
pip>=25.2
setuptools>=75.1.0
wheel

agora-token-builder==1.0.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.7
django-allauth>=0.63.3   # updated for Python 3.13
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
oauthlib==3.1.1
Pillow==9.0.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.4.0
pyOpenSSL==20.0.1
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
tzdata==2022.1
urllib3==1.26.6
zope.interface==5.4.0
